+++ changelog.md
@@ -4,3 +4,4 @@ Initial release
 ### 1.0.1
 - Beetroot stew may now be crafted using red mushrooms.
 - The default duration of the bonuses was increased to 3 days. This align with the apparition of phantoms
+- Gamerules from the mod now have their own category.